PCE: Fixed out-of-bound accesses in save state code
Timer variables are defined as SVAR_4 but they are 16bits, causing out-of-bound accesses during save/load.

It might be more appropriate to have reload/counter/running as uint8_t, but this break save state format so it will wait until we refactor save state code...

The size doesn't affect behavior anyway, we constrain the values and we rely on underflow arithmetic which is the same for 8/16/32.

+++ pce-go/components/pce-go/pce.h
@@ -148,10 +148,10 @@ typedef struct {
 	// Timer
 	struct {
 		uint32_t cycles_per_line;
-		uint16_t cycles_counter;
-		uint16_t counter;
-		uint16_t reload;
-		uint16_t running;
+		uint32_t cycles_counter;
+		uint32_t counter;
+		uint32_t reload;
+		uint32_t running;
 	} Timer;
 
 	// Joypad
